web/backend/pyproject.toml
[project]
name = "rapid-web"
dynamic = ["version"]
description = "RAPID Mission Control web service"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135,<1.0",
    "sqlmodel>=0.0.37,<1.0",
    "alembic>=1.18,<2.0",
    "pydantic-settings>=2.13,<3.0",
    "python-json-logger>=3.0,<4.0",
    "uvicorn[standard]>=0.34,<1.0",
    "watchdog>=6.0,<7.0",
    "tree-sitter>=0.24,<1.0",
    "tree-sitter-python>=0.23,<1.0",
    "tree-sitter-javascript>=0.23,<1.0",
    "tree-sitter-typescript>=0.23,<1.0",
    "tree-sitter-go>=0.23,<1.0",
    "tree-sitter-rust>=0.23,<1.0",
    "claude-agent-sdk>=0.1.59,<0.2",
    "sse-starlette>=2.0,<3.0",
]

[project.scripts]
rapid-web = "app.main:cli_entry"

[dependency-groups]
dev = ["pytest>=8.0", "httpx>=0.28", "pytest-asyncio>=0.25", "ruff>=0.11"]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.setuptools.dynamic]
version = {attr = "app.__version__"}

[tool.ruff]
line-length = 100
target-version = "py312"
